tools/requirements.txt
# Voice sample preparation dependencies (optional)
demucs
openai-whisper
soundfile
numpy
